+++ src/DateRangePicker/types.ts
@@ -1,0 +1,8 @@
+import { Dayjs } from "dayjs";
+
+export type DateRange = {
+  startDate: Dayjs | null;
+  endDate: Dayjs | null;
+};
+
+export type ClickState = "start" | "end";
